Added multiple swap views and moved to different folders
