+++ wdio/screen-objects/testIDs/Screens/QuoteView.js
@@ -1,0 +1,7 @@
+export const SWAP_SOURCE_TOKEN = 'select-source-token'
+export const SWAP_DEST_TOKEN = 'select-dest-token';
+export const SWAP_MAX_SLIPPAGE = 'swap_max-slippage';
+export const SWAP_SEARCH_TOKEN = 'swaps-search-token';
+
+
+
