+++ wdio/screen-objects/testIDs/Screens/SwapView.js
@@ -1,12 +1,7 @@
-export const SWAP_SOURCE_TOKEN = 'select-source-token'
-export const SWAP_DEST_TOKEN = 'select-dest-token';
-export const SWAP_MAX_SLIPPAGE = 'swap_max-slippage';
-
 export const SWIPE_TO_SWAP_BUTTON = 'swipe-to-swap-button';
 export const SWAP_QUOTE_SUMMARY = 'swap-quote-summary';
-
 export const SWAP_GAS_FEE = 'swap-gas-fee';
-export const SWAP_SEARCH_TOKEN = 'swaps-search-token';
+
